Split values into individual files
