+++ docs/index.md
@@ -6,7 +6,9 @@
 
 [:fontawesome-solid-download: Download from Modrinth](https://modrinth.com/mod/discarpet){ .md-button .md-button--primary }
 
-:fontawesome-brands-discord: If you need help, feel free to message me on Discord: `replaceitem#9118` or ping me in the [Carpet mod Discord](https://discord.gg/gn99m4QRY4)
+!!! help "Need help?"
+    If you need help, feel free to message me on Discord: `replaceitem#9118`
+    or ping me in the [Carpet mod Discord](https://discord.gg/gn99m4QRY4)
 
 [:fontawesome-solid-book: Getting started](./setup){ .md-button style="width: 100%" }
