+++ docs/values/commands/user-context-menu.md
@@ -1,0 +1,8 @@
+`dc_user_context_menu`
+
+| Property             | Type            | Description                                                                                                                                                                     |
+|----------------------|-----------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| `id`                 | String          | Id of the command                                                                                                                                                               |
+| `name`               | String          | Name of the command                                                                                                                                                             |
+| `description`        | String          | Description of the command                                                                                                                                                      |
+| `server`             | Server          | The server this application command is in, or null if it's a global application command                                                                                         |
