+++ src/main/java/Discarpet/script/events/DiscordEvents.java
@@ -3,15 +3,15 @@ package Discarpet.script.events;
 import Discarpet.Discarpet;
 import Discarpet.config.Bot;
 import Discarpet.mixins.CallbackListAccessor;
-import Discarpet.script.values.interaction.ButtonInteractionValue;
+import Discarpet.script.values.interactions.ButtonInteractionValue;
 import Discarpet.script.values.MessageValue;
-import Discarpet.script.values.interaction.MessageContextMenuInteractionValue;
-import Discarpet.script.values.interaction.ModalInteractionValue;
+import Discarpet.script.values.interactions.MessageContextMenuInteractionValue;
+import Discarpet.script.values.interactions.ModalInteractionValue;
 import Discarpet.script.values.ReactionValue;
-import Discarpet.script.values.interaction.SelectMenuInteractionValue;
-import Discarpet.script.values.interaction.SlashCommandInteractionValue;
+import Discarpet.script.values.interactions.SelectMenuInteractionValue;
+import Discarpet.script.values.interactions.SlashCommandInteractionValue;
 import Discarpet.script.values.UserValue;
-import Discarpet.script.values.interaction.UserContextMenuInteractionValue;
+import Discarpet.script.values.interactions.UserContextMenuInteractionValue;
 import carpet.CarpetServer;
 import carpet.script.CarpetEventServer.Event;
 import carpet.script.value.BooleanValue;
