+++ src/main/java/Discarpet/script/util/Registration.java
@@ -17,43 +17,43 @@ import Discarpet.script.parsable.parsables.EmbedAuthorParsable;
 import Discarpet.script.parsable.parsables.EmbedFieldParsable;
 import Discarpet.script.parsable.parsables.EmbedFooterParsable;
 import Discarpet.script.parsable.parsables.EmbedParsable;
-import Discarpet.script.parsable.parsables.MessageContextMenuBuilderParsable;
+import Discarpet.script.parsable.parsables.commands.MessageContextMenuBuilderParsable;
 import Discarpet.script.parsable.parsables.TimestampParsable;
 import Discarpet.script.parsable.parsables.MessageContentParsable;
 import Discarpet.script.parsable.parsables.ModalParsable;
 import Discarpet.script.parsable.parsables.SelectMenuOptionParsable;
 import Discarpet.script.parsable.parsables.SelectMenuParsable;
-import Discarpet.script.parsable.parsables.SlashCommandBuilderParsable;
-import Discarpet.script.parsable.parsables.SlashCommandOptionChoiceParsable;
-import Discarpet.script.parsable.parsables.SlashCommandOptionParsable;
+import Discarpet.script.parsable.parsables.commands.SlashCommandBuilderParsable;
+import Discarpet.script.parsable.parsables.commands.SlashCommandOptionChoiceParsable;
+import Discarpet.script.parsable.parsables.commands.SlashCommandOptionParsable;
 import Discarpet.script.parsable.parsables.TextInputParsable;
-import Discarpet.script.parsable.parsables.UserContextMenuBuilderParsable;
+import Discarpet.script.parsable.parsables.commands.UserContextMenuBuilderParsable;
 import Discarpet.script.parsable.parsables.WebhookMessageProfileParsable;
 import Discarpet.script.parsable.parsables.WebhookProfileParsable;
 import Discarpet.script.parsable.parsables.WebhookProfileUpdaterParsable;
 import Discarpet.script.values.AttachmentValue;
-import Discarpet.script.values.command.MessageContextMenuValue;
-import Discarpet.script.values.command.SlashCommandValue;
-import Discarpet.script.values.command.UserContextMenuValue;
-import Discarpet.script.values.interaction.ButtonInteractionValue;
+import Discarpet.script.values.commands.MessageContextMenuValue;
+import Discarpet.script.values.commands.SlashCommandValue;
+import Discarpet.script.values.commands.UserContextMenuValue;
+import Discarpet.script.values.interactions.ButtonInteractionValue;
 import Discarpet.script.values.ChannelValue;
 import Discarpet.script.values.EmojiValue;
 import Discarpet.script.values.MessageValue;
-import Discarpet.script.values.interaction.MessageContextMenuInteractionValue;
-import Discarpet.script.values.interaction.ModalInteractionValue;
+import Discarpet.script.values.interactions.MessageContextMenuInteractionValue;
+import Discarpet.script.values.interactions.ModalInteractionValue;
 import Discarpet.script.values.ReactionValue;
 import Discarpet.script.values.RoleValue;
-import Discarpet.script.values.interaction.SelectMenuInteractionValue;
+import Discarpet.script.values.interactions.SelectMenuInteractionValue;
 import Discarpet.script.values.ServerValue;
-import Discarpet.script.values.interaction.SlashCommandInteractionOptionValue;
-import Discarpet.script.values.interaction.SlashCommandInteractionValue;
-import Discarpet.script.values.interaction.ApplicationCommandInteractionValue;
+import Discarpet.script.values.interactions.SlashCommandInteractionOptionValue;
+import Discarpet.script.values.interactions.SlashCommandInteractionValue;
+import Discarpet.script.values.interactions.ApplicationCommandInteractionValue;
 import Discarpet.script.values.UserValue;
 import Discarpet.script.values.WebhookValue;
 import Discarpet.script.values.common.DiscordValue;
-import Discarpet.script.values.interaction.InteractionValue;
+import Discarpet.script.values.interactions.InteractionValue;
 import Discarpet.script.values.common.MessageableValue;
-import Discarpet.script.values.interaction.UserContextMenuInteractionValue;
+import Discarpet.script.values.interactions.UserContextMenuInteractionValue;
 import carpet.script.annotation.AnnotationParser;
 import carpet.script.annotation.OutputConverter;
 import carpet.script.annotation.SimpleTypeConverter;
