+++ src/main/java/Discarpet/script/values/interactions/ApplicationCommandInteractionValue.java
@@ -1,4 +1,4 @@
-package Discarpet.script.values.interaction;
+package Discarpet.script.values.interactions;
 
 import carpet.script.value.StringValue;
 import carpet.script.value.Value;
