+++ src/main/java/Discarpet/script/values/interactions/SlashCommandInteractionValue.java
@@ -1,4 +1,4 @@
-package Discarpet.script.values.interaction;
+package Discarpet.script.values.interactions;
 
 import carpet.script.value.ListValue;
 import carpet.script.value.MapValue;
